[Setup] Install dependencies early in Dockerfile
Modified the PyAleph Dockerfile to separate the install of
the dependencies from PyAleph itself. This way, we can install
the requirements once and cache them in a layer of the image,
avoiding to fetch them each time.

A new script, `extract_requirements.py`, is used to read
the install requirements from setup.cfg into a temporary
requirements.txt file.

Removed the install of pytest from the Docker image as it is
not relevant to a Docker image.