[P2P] Added tests and fixed issues
Added tests for the P2P service. We now test the connection and
pubsub functionalities.

The new tests spawn JS P2P daemons and act as simplified nodes
exchanging messages. We test the following features:

* connection to a remote node
* exchange of "hello" messages
* subscription and publication of messages on P2P topics.

The tests for the P2P streaming protocol will follow after a bug
within p2pclient and/or the P2P daemon is fixed.