Change: Sort pending messages by number of retries in processing
We observed that some messages fail being processed correctly many times. This could be due to temporary issues and the message should still be processed and not ignored.

In order to prevent these messages from obstructing other messages to be processed, this change sorts the pending messages during processing by the ascending number of retries, such that messages that have now or little been retried are processed first.

This should improve the responsiveness of the nodes when processing new messages.