[Doc] Update documentation for upcoming release
Updated the install guide to reflect the addition of the P2P
daemon and the modification of the key generation procedure.

Moved the configuration section of the Docker Compose section
to the start of the install guide as it is generic.

Moved the node synchronisation section out of the guides as
it is not referenced there.

Updated the changelog.