[Storage] Rewrite status codes as exceptions
Simplified the storage code and made it more pythonic by replacing
status codes like None and -1 by exceptions. Two new exceptions
describe if the content the user is looking for is invalid (-1)
or just currently unavailable because of a synchronisation issue
or a technical issue (None).

This simplifies the whole logic and avoids the propagation of weird
codes all over the place.

Added new dataclasses to keep track of the retrieved content,
its hash and the source from which the data was fetched.
The `RawContent` class keeps track of binary content while
the `JsonContent` class represents raw content that was interpreted
as JSON.