[Storage] Use Aleph message model for validation
Mirrored the validation system used for forget messages in store
messages. Used the `StoreMessage` class to validate the message
and content dictionaries.

We now discard the message if it fails validation instead of
retrying it through the default exception handler. This fixes
an issue where invalid store messages would be retried indefinitely.

Added unit tests for the storage handler.