[ETH] Upgrade packer to EIP 1559 gas fees
Upgraded the packer task to use EIP 1559 fees instead of legacy
fees. The task now specifies the `maxFeePerGas` parameter to
specify the maximum allowed value for the base + priority fee.
It also specifies a maximum priority fee of 2 gwei by default,
as this seems to be a recommended value to get the transaction
included by miners.