[Infra] Added a startup script
The new `run_aleph_ccn.sh` script is now used to start a CCN.
This script launches the CCN after verifying that the other
services are already up. We use the open source `wait-for-it`
tool to monitor the other services.

Added a new script to read configuration values from a shell
script.

Modified the Docker file to use `entrypoint` instead of `cmd`.