[Processing] Simplify incoming message functions (#227)
Improved the `incoming` function to simplify it and make it more
generic.

Removed the `bulk_operation` flag as we always wish for bulk DB
operations. The function now returns a status + a list of DB
operations. DB operations are a new class that contains a Mongo
operation and the name of the collection to target.
This avoids passing lists as arguments and modifying them in place.

The loop to await individual message tasks now waits for all tasks
to complete instead of failing on the first operation.
This fixes an issue where, if an exception occurs soon enough in
the batch, the output of still-running tasks could be lost.

Added type hints.