[Messages] Ignore confirmations in capped collection
Fixed an issue where the message processing job would fail while
writing a confirmation to the capped messages collection.

In MongoDB, it is not possible to increase the size of a document
in a capped collection. As this particular collection is only used
for the websockets to advertise new messages, we now simply
ignore the confirmation for this collection.

Added a unit test to test the confirmation flow.