[Tests] Make HTTP endpoints testable
It is now possible to write API tests with the test client of aiohttp.
Added an example test that checks that the /version endpoint returns
the expected version.

Refactored the declaration of endpoints in the initialization of
the app. We now have a `register_routes` function centralized in
a single module. All references to the global app object inside the
implementation of endpoints were replaced by accesses to `request.app`.
This allows to import the app object, configured properly, without
importing other files.