[Permissions] Process unauthorized messages gracefully
Fixed an issue where the check of permissions would fail if
the content.address field of the message was set to a different
address than the sender and the owner of this address did not
authorize the sender to post message on their behalf.

Added basic unit tests for the permission system.