[Storage] Garbage collector CLI
Added a new command to the CCN operator CLI to run a garbage
collector on local storage. The new `gc run` command lists
all the files that are not linked to any message or permanent
pin and deletes them.

Using the --verbose option, the command will print more details
on the files it will preserve and delete. The --dry-run option
allows to run the GC without actually deleting any file.