[DB] Use unordered bulk inserts
We now use unordered bulk inserts when processing the results
of the pending txs and messages jobs. Unordered writes guarantee
that all the writes will be attempted even if errors occur.

This property help fixing a long-standing issue with capped
collections. Given the design of the messages job, it is possible
that the job will handle the same message several times in parallel,
resulting in multiple updates to the same document.
This is problematic for the `CappedMessage` collection as capped
collections do not support updates that increase the size of
a document.

We now make the assumption that we do not care if bulk writes
to capped collections fail, and the unordered write guarantees
that we do not fail to insert a message because an error occurred.

Introduced two new context managers in the toolkit module:
* the new `Timer` class allows to measure the time needed to run
  some operations
* `ignore_exceptions` allows to ignore one or more exceptions
  in a block of code.