[Messages] Sync job
Added a new job that synchronizes unconfirmed messages across
the network. The goal of this job is to re-send messages missed
by nodes with the ability to push data on-chain. This can happen
because of various issues like server downtimes or bugs.

This job works in three parts:
* the publisher task periodically sends the list of all the messages
  older than the last TX block that have yet to be confirmed by
  on-chain data.
* the receiver task stores the list of unconfirmed messages for
  each peer detected on the network.
* the sync/aggregator task aggregates the confirmation data from
  all the nodes and fetches messages using the HTTP API.
  These messages are added to the pending message queue.

This solution is less expensive that constantly sharing all
the messages across all the nodes and guarantees that the network
will be synchronized eventually as long as the on-chain data
synchronization jobs are working. With the current implementation,
a message can remain out of sync at a maximum until a new TX
is published on-chain + the job period (5 minutes currently).