[Messages] Pydantic schema for message validation (#272)
Added new Pydantic models/schemas to validate Aleph messages
coming from the network. These models are similar to the ones
implemented in aleph-message, but focus on validating messages
before processing them.

This validation replaces the validation in check_message
and will later be used to use objects as early as possible
in the processing of messages.

Added tests for the major types of messages.