[Fix] type error in storage/forget handlers (#290)
Implemented a temporary fix to resolve a TypeError that sometimes
occurs in the forget/storage handler. When building Aleph message
content models, the content would sometimes be passed twice to
the constructor, once in the message itself and a second time
explicitly with the `content` key.

We fix this by setting the `content` key directly in the message.
This is a temporary fix until Pydantic models are used everywhere
in the pipeline.