[Forget] Forget aggregates (#289)
Added support for the new `aggregates` field in FORGET messages.
Using this field, users can now specify one or more aggregates
to forget. This will lead to the deletion of all the messages
linked to these aggregates.

For now, the implementation simply lists all the messages stored
on the node and deletes them.