[Fix] Set default item type early (#298)
Fix for the previous fix. Some messages end up with check_message
set to False, bypassing the validation in `incoming`. We now
set the default value of `item_type` as the first thing in
the function.

Note that this is a temporary fix that will disappear as soon
as we use Pydantic models everywhere.