[Messages] No blocking on end of queue
Fixed an issue where the pending message job would block on
the final messages in the queue and stop processing newer messages.

Once the job finishes the loop on all the messages in the pending
message collection, the previous implementation waits until all
the message tasks finish. This causes a delay of several hours
until the node finishes these tasks and is able to process newer
pending messages again. Messages end up being processed, but far
later than expected.

The issue arises because we never remove messages from the pending
queue if we fail to retrieve the associated content. The job
then always has messages in the queue, causing the issue.

Fixed the issue by allowing the loop to restart without waiting
for messages to be processed. We now compute an individual ID
for each pending message and add it to a set. The job will simply
ignore any message that is already being processed, allowing
for newer messages to be taken into account.