[Messages] Use Pydantic message models everywhere (#301)
Propagated the use of Pydantic message models to all the functions
related to message processing.

Extended the Pydantic message models by introducing a new
message class, `ValidatedBaseMessage`. This schema represents
fully-loaded messages, including their content field.

Split the message validation function, check_message, in two parts:
* `parse_message` checks that the message is semantically valid,
  i.e. that all the required fields are present, are of the correct
  type and have sensible values.
* `verify_signature` checks the signature of the message using
  the public key of the sender.

We now call parse_message as early as possible in the process.
This allows to simplify hypotheses about the presence/absence
of specific fields and allows to simplify the codebase.

Improved the typing of the message models by using generics
to represent the message and item type fields.

Split the message validation function, check_message, in two parts:
* `parse_message` checks that the message is semantically valid,
  i.e. that all the required fields are present, are of the correct
  type and have sensible values.
* `verify_signature` checks the signature of the message using
  the public key of the sender.

We now call parse_message as early as possible in the process.
This allows to simplify hypotheses about the presence/absence
of specific fields and allows to simplify the codebase.

Modified tests to use the message classes. Removed a few tests
that tested corner cases linked to incomplete message dictionaries.