Feature: use the file system to store files
This is the combination of the following commits:

- Feature: filesystem storage engine

Problem: benchmarks show that storing files on GridFS is slower
than storing them directly on the filesystem.

Solution: implement a storage engine class that manipulates files
on the filesystem directly.

- Internal: remove IPFS API globals

Problem: the IPFS API client is a global variable. As demonstrated
in previous PRs, we do not like global variables.

Solution: introduce a new service for IPFS-related functions and
instantiate the API client without globals.

- fix data folder permissions

- Fix: increase timeout on pending TX chain data file fetch

Problem: fetching chain data files while processing pending TXs
may time out.

Solution: increase the timeout from 10 to 60 seconds.

We expect chain data files to always be available, so there is
no clear definition of what should be done if the operation really
fails. For the moment, we accept this situation, just increase
the timeout to reduce the amount of exceptions logged (10 seconds
was too short) and let the handler retry, forever if needed.

Fix: stop pushing historical messages on websocket

Problem: bulk inserts on the capped collection used to feed
the websocket are slow and unnecessary as getting bombarded
by 1000s of messages at sync does not make sense.

Solution: stop inserting on-chain messages to the capped
message collection.