Fix: general 500 error on posts endpoint
Problem: the "/api/v0/posts.json" endpoint returns a 500 error
because of an invalid parameter being used in a function
(an attempt to remove the MongoDB object ID from the endpoint).

Solution: remove the object ID directly in the aggregate query.

Added a simple test to verify that the endpoint is at least working
a little bit.