Feature: exponential message retry times
Problem: the current implementation of message retries may not leave
enough time for the situation that cause the message processing failure
to resolve itself, resulting in wrongly discarded messages.

Solution: implement exponential retry times. With exponential retry
times, messages are retried less times (10 times instead of 1000
in the default configuration) and are guaranteed to remain ~15
minutes in the processing queue.