Internal: test error code enum values
Problem: error codes from the `ErrorCode` enum must be mapped in
the `error_codes` table to avoid a series of issues with rejected
messages. This condition is not guaranteed by the code itself.

Solution: add a unit test to guarantee that migration scripts
insert the correct values in `error_codes`.