Feature: /addresses/{address}/files endpoint
Problem: account.aleph.im relies on the /messages endpoint to
get the files stored by an address. This feature is not supported
anymore as it relied on modifying the content field of messages to
add fields, which we want to avoid moving forward.

Solution: Add a new endpoint to retrieve file metadata for an address.
/addresses/{address}/files returns the total amount of files, their
total size and a paginated list of files.

* Added a not null hash index on `file_pins.owner` to accelerate queries
  with the new endpoint.
* Migrated the balance endpoint to /addresses/{address}/balance for
  consistency.