Feature: fetch message/sync events from the Aleph indexer
Problem: we wish to use the Aleph message smart contract to create
message and sync events. We then wish to process these events
on the CCNs as Aleph messages.

Solution: use the Aleph smart contract indexer to retrieve message
and sync events for Binance Smart Chain and Ethereum.

The new `AlephIndexerReader` class allows to read any event from
the smart contract. The state of the synchronization between the CCN
and the indexer is maintained in the new `indexer_sync_status` table.