Fix: wait for P2P service at startup
Problem: some users experience an issue where the node crashes
at startup on the first attempt to connect to the P2P service.
This occurs because the P2P service takes too long to start its
HTTP server, usually because one of the bootstrap peers is too slow
to respond or because of a configuration issue.

Solution: wait for the P2P service HTTP server to be available in
the node startup script.