Fix: restrict storage endpoints to files stored on aleph.im
Problem: the `/storage/` and `/storage/raw/` endpoints do not
check if the requested file belongs to the aleph.im network,
whether it's the file associated to a STORE message, the content
of a non-inline message or a tx sync file.
For IPFS CIDs, this can cause the node to fetch any file on
the IPFS network, causing additional network load for an unintended
usage.

Solution: check if any file pin exists for the requested file.
If it is not the case, return a 404.