Fix: return 413 on /storage/ for files larger than 100MB
Problem: the /storage/ endpoints return a 404 for IPFS files that
are over 100MB. This is unclear for developers as they expect the
file to be present but cannot download it.

Solution: send a 413 - Payload Too Large code instead. We now check
the existence and size of the file as soon as possible. This will
also reduce the load on the IPFS daemon as the previous implementation
would always request the first 100MB of the file even if the file
is too large.