Fix: aiohttp assert exception on connection lost (#406)
Problem: a bug in aiohttp causes an assertion error to appear in
the logs whenever an HTTP connection is lost.

Solution: there is no mainline fix in aiohttp. It is not trivial
to install a forked version of aiohttp, so the simplest solution
for now is to apply the aiohttp patch directly.

Bumped aiohttp to 3.8.4 at the same time.