Fix: uploaded files are not downloadable
Problem: when a user uploads a file with the /storage and /ipfs
add_file/add_json endpoints, the corresponding files are not
downloadable until a corresponding STORE message is processed.
This is problematic if the message is not sent to the same node
or if the message fails to be processed because of a bug. The file
would then be impossible to download, making the message unprocessable
for other nodes.

Solution: add an entry in the `files` table for each uploaded file.