Fix: avoid channel closed exceptions in message websocket
Problem: upon closing, the message websocket may attempt to close
the RabbitMQ channel. This appears to be caused by the handling
of `CancellationError` in the `close()` method of the queue
iterator.

Solution: replace the `iterator` implementation with a simpler
`consume` + callback implementation. This guarantees that
the secondary task will never attempt to close the channel.
Replace task cancellation by a call to `queue.cancel()` as
aiormq handles cancellation exceptions by closing the channel.