Feature: process instance messages (#443)
Problem: users cannot send `INSTANCE` messages because the format
is not yet supported by CCNs.

Solution: support them, obviously. Modified the program handler
to support both programs and instances. Renamed most program-related
functions to the more generic `vm` naming convention.

DB-wise, the same tables are reused for instances and programs.
The `vms` table (previously named `programs`) now contains entries
for instances and programs. A new table is dedicated to rootfs
volumes as that is an instance-only feature. The advantage of reusing
the same tables is that the logic for instances and programs can
be unified, for example the balance/cost views only require minor
updates.