Fix: no DB calls in message websocket
Problem: the websocket uses the DB to apply filters to processed
messages to determine whether they must be sent on the websocket.
This leads to an exhaustion of DB connections when many websockets
are open with the same filters.

Solution: send the full message on the RabbitMQ topic and apply
filters in Python instead of using a DB query.