Fix: prevent cancellation of the message websocket
Problem: resources are not cleaned up properly when the message
websocket handler is cancelled. This is because aiohttp appears
to raise a CancellationError.

Solution: use `asyncio.shield()` on the whole handler.
Added a decorator to avoid splitting the implementation of shielded
functions.
Protected the POST /messages endpoint as well as it also uses
a message queue.

fix