Internal: split chain service (#487)
Problem: the chain service is in charge of providing all chain-related
functionalities, namely: verify signature and read/write on chain. It
turns out that most use cases of the chain service only care about
signature verification, meaning that the dependencies required for the
chain service are not necessary most of the time. As we plan on adding
more dependencies for chain readers, the chain service is getting too
complex for its own good.

Solution: split the chain service into a signature verifier class and a
chain connector class. This way, the signature verifier can be
instantiated without dependencies in all the places where it is required
and the chain connector can get a bit more complex without impacting the
rest.

Other changes:
* Removed the conditional import of chain connectors as they must all be
  installed on each node.