Fix: All node metrics were returned by default
This causes a lot of data to be returned in the case of old nodes and is usually unnecessary.

Solution: Default to limiting the query to the last 2 weeks of data, or to 2 weeks before the specified date of last data.

This also renames `start_date -> start_timestamp` and `end_date -> end_timestamp` to prevent confusion since these fields are expected to be timestamps and note datetime or ISO 8601.